requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.1.6
django-extensions==3.2.3
django-filter==25.1
django-import-export==4.3.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
Faker==36.1.1
geographiclib==2.0
geopy==2.4.1
httpie==3.2.4
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
pillow==11.1.0
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.10.1
PySocks==1.7.1
python-decouple==3.8
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
setuptools==75.8.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.1
urllib3==2.3.0
